refactor: rename queryBuilder to MetricQueryBuilder
